mcps/exchange-rate/pyproject.toml
[project]
name = "exchange-rate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.27.0",
    "python-decouple>=3.8",
]
